parameterized 3 methods-email,password and submit-so values can be entered for diff cases
