+++ src/test/java/BaseTest.java
@@ -1,9 +1,8 @@
 import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.firefox.FirefoxDriver;
-import org.openqa.selenium.safari.SafariDriver;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.BeforeSuite;
@@ -13,12 +12,11 @@ import java.time.Duration;
 
 public class BaseTest {
 
+    public static WebDriver driver = null;
     @BeforeSuite
     static void setupClass() {
         WebDriverManager.chromedriver().setup();
     }
-
-    public static WebDriver driver = null;
 
     @BeforeMethod
     public static void launchBrowser() {
@@ -26,11 +24,24 @@ public class BaseTest {
         driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
     }
     @AfterMethod
-    public static closeBrowser(){
+    public static void closeBrowser(){
         driver.quit();
     }
 
+    public static void clickSubmit() {
+        WebElement submitButton = driver.findElement(By.cssSelector("button[type=submit]"));
+        submitButton.click();
+    }
 
+    public static void provideEmail(String email) {
+        WebElement emailField = driver.findElement(By.cssSelector("[type='email']"));
+        emailField.sendKeys(email);
+    }
+
+    public static void providePassword(String password) {
+        WebElement passwordField = driver.findElement(By.cssSelector("[type ='password']"));
+        passwordField.sendKeys(password);
+    }
 }
