+++ src/test/java/LoginTests.java
@@ -10,8 +10,6 @@ public class LoginTests extends BaseTest {
     @Test
     public static void loginEmptyEmailPasswordTest () {
 
-        WebDriver driver = new ChromeDriver();
-        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
         //update url
         String url = "https://testpro.io/";
         driver.get(url);
